# public-docs/openapi.yaml
openapi: 3.0.3
info:
  title: PKV Sync API
  version: 1.0.10
  license:
    name: AGPL-3.0-only
  contact:
    name: cyberkurry
  description: |
    HTTP API for PKV Sync clients and admin integrations.

    All `/api/*` routes are protected by the deployment key middleware and the
    plugin User-Agent guard in the production server stack. Authenticated user
    and admin routes additionally require a bearer device token. Production
    responses include clickjacking, MIME sniffing, referrer, and CSP headers;
    HSTS is emitted when `public_host` is configured.

    Starting at v1.0.0, this OpenAPI document is the public REST contract.
    Routes outside this document, including Admin Web UI form handlers, are
    internal implementation details unless documented separately.
servers:
  - url: https://sync.example.com
security:
  - DeploymentKey: []
    PluginUserAgent: []
...
components:
  securitySchemes:
    DeploymentKey:
      type: apiKey
      in: header
      name: X-PKVSync-Deployment-Key
      description: Deployment key distributed through the server share URL.
    PluginUserAgent:
      type: apiKey
      in: header
      name: User-Agent
      description: Must use the PKV Sync plugin User-Agent format accepted by the server.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: pks_token
  parameters:
    VaultId:
      in: path
      name: vault_id
      required: true
      schema:
        type: string
      description: Vault UUID-like id without dashes.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing, invalid, revoked, or expired bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Resource not found, or deployment key/User-Agent rejected by the production stack
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
paths:
...
  /git/{vault_id}/git-upload-pack:
    post:
      tags: [Git]
      summary: Git smart HTTP — stateless RPC for upload-pack
      description: |
        Stateless-RPC body of a `git clone` / `git fetch` request. The
        request body carries the want/have negotiation; the response is
        the packfile. Body is capped at 10 MiB; legitimate negotiation
        requests are far smaller.
      parameters:
        - $ref: '#/components/parameters/VaultId'
      requestBody:
        required: true
        content:
          application/x-git-upload-pack-request:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Pack response (`application/x-git-upload-pack-result`)
          content:
            application/x-git-upload-pack-result:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Git smart HTTP is disabled or `git` is not available
...